fix: Move SafeTranslatorUtility to Core (from Editor) to fit with new asmdef files surrounding test assemblies.
