+++ Assets/Plugins/Source/Core/Utility/SafeTranslatorUtility.cs.meta
