--- Assets/Tests/ClientTests.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dc964976d7894944588b101b72e3a8bb
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
