--- Assets/Tests/IntegrationTests.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: af13f1cd10ae3c94a8eedc252362bea9
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
