+++ Assets/Tests/PlayMode/AchievementsTests.cs
@@ -33,7 +33,7 @@ namespace PlayEveryWare.EpicOnlineServices.Tests.IntegrationTests
     /// <summary>
     /// Integration tests for achievements.
     /// </summary>
-    public class EOSAchievementTests : EOSTestBase
+    public class AchievementsTests : EOSTestBase
     {
         /// <summary>
         /// Checks to see if we can get a count for the number of achievements for the product user.
