+++ Assets/Tests/PlayMode/com.playeveryware.eos.tests.playmode.asmdef
@@ -2,10 +2,11 @@
     "name": "com.playeveryware.eos.tests.playmode",
     "rootNamespace": "",
     "references": [
-        "UnityEngine.TestRunner",
-        "UnityEditor.TestRunner",
-        "com.playeveryware.eos-Editor",
-        "com.playeveryware.eos.core"
+        "GUID:793f42bcdc73d104e91ec29d214254da",
+        "GUID:be1f6e9efffcfce41aeb5c2f4c27cf75",
+        "GUID:3a63500f0eef43a438c0553491f7c5da",
+        "GUID:0acc523941302664db1f4e527237feb3",
+        "GUID:27619889b8ba8c24980f49ee34dbb44a"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
@@ -16,7 +17,8 @@
     ],
     "autoReferenced": false,
     "defineConstraints": [
-        "UNITY_INCLUDE_TESTS"
+        "UNITY_INCLUDE_TESTS",
+        "!EOS_DISABLE"
     ],
     "versionDefines": [],
     "noEngineReferences": false
