--- Assets/Tests/SceneScripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dc7f8a9da80a4444db3ca092153c2fd6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
