--- Assets/Tests/TestCategories.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f11a7844cf4969b4da6ce1c4b8d352a6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
