--- Assets/Tests/WaitUntilDone.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: db8c86cc57f5e3441b9f890b2fbf5ef2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
